Added the class HWObjActuatorCommand to be able to handle Actuator Hardware Objects in beamline actions.
